feat(v1.18.0): API enhancements, exports, security banner, PHPStan level 7
API (#231, #232, #258, #260):
- Subnet GET: site_id in response; ?counts=1 adds address_counts object
- Addresses GET: ?site_id= filter
- New resource=unassigned endpoint (IPv4, paginated, max /24)

Exports (#227, #259, #261):
- export_address_history.php: per-address change history CSV
- export_subnet_utilization.php: utilization summary CSV
- export_addresses.php: cross-subnet all-addresses export when no subnet_id

UI/UX (#233, #234, #239, #240, #241, #242):
- Audit log: Client IP column displayed
- status.php: schema_version in health-check JSON
- address_history.php: styled error page on missing address_id
- addresses.php: empty-state nav link when no subnet selected
- Dismissible security banners on db_tools, import_csv, demo_gate
- DHCP Pools moved to Admin dropdown in nav

PHPStan level 6 → 7: fixed string|false (inet_pton, json_encode, unpack),
PDOStatement|false (?: throw pattern), list<T> narrowing (array_values),
and int casts for strtotime() across lib.php, api.php, status.php,
address_history.php, import_csv.php, subnets.php, audit.php

Tests: API suite 58→65, CDP suite 80→94

+++ Simple-PHP-IPAM/audit.php
@@ -43,7 +43,7 @@ if ($filterFrom !== '') {
 }
 if ($filterTo !== '') {
     $wheres[] = 'created_at < :to';
-    $params[':to'] = date('Y-m-d', strtotime($filterTo . ' +1 day')) . ' 00:00:00';
+    $params[':to'] = date('Y-m-d', (int)strtotime($filterTo . ' +1 day')) . ' 00:00:00';
 }
 $where = $wheres ? 'WHERE ' . implode(' AND ', $wheres) : '';
